[zephyr] Move SubprocessRunner child entry to its own module
SubprocessRunner spawned each shard as `python -m zephyr.runners`, but
zephyr/__init__.py imports zephyr.runners transitively (__init__ ->
execution -> runners). runpy therefore found zephyr.runners already in
sys.modules before executing it as __main__, emitting:

  RuntimeWarning: 'zephyr.runners' found in sys.modules after import of
  package 'zephyr', but prior to execution of 'zephyr.runners'

and re-running the module body a second time in the child.

Split the child-process entry point (the periodic counter/status helpers,
_execute_shard_subprocess, _subprocess_main, and the __main__ guard) into
a new leaf module zephyr/_shard_subprocess.py and point SubprocessRunner at
`python -m zephyr._shard_subprocess`. Nothing in the package-init graph
imports that module, so runpy no longer finds it pre-loaded. runners.py
keeps the runner classes and the shared in-process helpers
(_InProcessWorkerContext, _run_stage_with_ctx).

The broader cleanup of zephyr's eager __init__ re-export aliasing (~93
consumers) is left as a separate follow-up; see
.agents/projects/zephyr-import-cleanup.md.
